# openapi/webhooks.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Gryt webhooks",
    "version": "1",
    "description": "Post messages into a Gryt channel from anywhere that can send an HTTP request, and manage the webhooks that do it. Every path is relative to your Gryt server.",
    "license": {
      "name": "AGPL-3.0-or-later",
      "identifier": "AGPL-3.0-or-later"
    }
  },
  "servers": [
    {
      "url": "https://{server}",
      "variables": {
        "server": {
          "default": "gryt.example.com",
          "description": "Your Gryt server's address."
        }
      }
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "A member's access token. Needs the manage_webhooks permission."
      }
    },
    "schemas": {
...
      "WebhookCreate": {
        "type": "object",
        "properties": {
          "channel_id": {
            "default": "",
            "description": "The channel the webhook posts to.",
            "type": "string"
          },
          "display_name": {
            "description": "Cut to 64 characters. Defaults to \"Webhook\".",
            "type": "string"
          },
          "avatar_file_id": {
            "description": "An uploaded file you can read.",
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "InvalidPayload": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "const": "invalid_payload"
          },
          "message": {
            "type": "string"
          },
          "problems": {
            "maxItems": 20,
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string",
                  "description": "Where in the payload, like `cards[0].fields[3].value`."
                },
                "code": {
                  "type": "string",
                  "enum": [
                    "required",
                    "wrong_type",
                    "too_long",
                    "too_many",
                    "invalid_url",
                    "invalid_color",
                    "invalid_timestamp",
                    "empty_card",
                    "total_too_long",
                    "invalid_format",
                    "invalid"
                  ]
                },
                "limit": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                },
                "message": {
                  "type": "string"
                }
              },
              "required": [
                "path",
                "code",
                "message"
              ]
            }
          }
        },
        "required": [
          "error",
          "message",
          "problems"
        ]
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
      "Webhook": {
        "type": "object",
        "properties": {
          "webhook_id": {
            "type": "string"
          },
          "channel_id": {
            "type": "string"
          },
          "display_name": {
            "type": "string"
          },
          "avatar_file_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "token": {
            "type": "string"
          },
          "created_by_server_user_id": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        },
        "required": [
          "webhook_id",
          "channel_id",
          "display_name",
          "avatar_file_id",
          "token",
          "created_by_server_user_id",
          "created_at",
          "updated_at"
        ]
      }
    }
  },
  "paths": {
...
    "/api/webhooks": {
      "get": {
        "operationId": "listWebhooks",
        "summary": "List webhooks",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "The webhooks in channels you can see.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Webhook"
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "No manage_webhooks permission.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createWebhook",
        "summary": "Create a webhook",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created. `url` is what to post to.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Webhook"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "url": {
                          "type": "string"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The body was refused, or isn't valid JSON (`invalid_json`).",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/InvalidPayload"
                    },
                    {
                      "$ref": "#/components/schemas/Error"
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "description": "No manage_webhooks permission.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "The avatar file doesn't exist or you can't read it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "The body is over 100 KB (`body_too_large`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}